New casting infrastructure in the core, use smartCAST!

+++ code/runtime/include/impl/irt_loop_sched.impl.h
@@ -177,7 +177,7 @@ inline static void irt_schedule_loop_guided_chunked(irt_work_item* self, uint32 
 		uint64 bsize = sched_data->block_size;
 		if(irt_atomic_bool_compare_and_swap(&sched_data->completed, comp, comp+bsize, uint64)) {
 			uint64 new_bsize = (uint64)((bsize / base_range.step) * 0.8) * base_range.step;
-			new_bsize = MAX(new_bsize, sched_data->policy.param.chunk_size * base_range.step);
+			new_bsize = MAX(new_bsize, (uint64)(sched_data->policy.param.chunk_size * base_range.step));
 			irt_atomic_bool_compare_and_swap(&sched_data->block_size, bsize, new_bsize, uint64);
 			base_range.begin = comp;
 			base_range.end = MIN(comp+bsize, final);
